+++ lang/en/chats.php
@@ -1,23 +1,21 @@
 <?php
 
-
-
 return [
 
     /**-------------------------
      * Chats
      *------------------------*/
     'labels' => [
-        'heading'=>'Chats',
-        'you'=>'You',
-  
-    ],
-        
-        'inputs' => [
-            'search' => [
-                'label' => 'Search Conversations',
-                'placeholder' => 'Search',
-            ],
-        ],
-    ];
+        'heading' => 'Chats',
+        'no_conversations_yet' => 'No conversations yet',
+        'you' => 'You',
 
+    ],
+
+    'inputs' => [
+        'search' => [
+            'label' => 'Search Conversations',
+            'placeholder' => 'Search',
+        ],
+    ],
+];