+++ resources/views/livewire/widgets/wire-chat.blade.php
@@ -217,9 +217,7 @@
 
             <div  x-show="!show && !chatIsOpen " class="m-auto  justify-center flex gap-3 flex-col  items-center ">
 
-                <h4
-                    class="font-medium p-2 px-3 rounded-full font-semibold bg-gray-50 dark:bg-gray-800 dark:text-white dark:font-normal">
-                    Select a conversation to start messaging</h4>
+                <h4 class="font-medium p-2 px-3 rounded-full font-semibold bg-gray-50 dark:bg-gray-800 dark:text-white dark:font-normal">@lang('wirechat::widgets.wirechat.messages.welcome')</h4>
   
             </div>